- Changes to make LMbench better handle large memory systems   (e.g., use size_t as appropriate, and so forth)
(Logical change 1.85)


git-svn-id: https://svn.code.sf.net/p/lmbench/code/trunk@911 9cc621ae-7264-4df2-8f7e-c588375cf588

+++ lmbench2/src/lat_mmap.c
@@ -28,7 +28,7 @@ char	*id = "$Id$\n";
  * mappings reported.
  */
 void
-mapit(int fd, int size, int random)
+mapit(int fd, size_t size, int random)
 {
 	char	*p, *where, *end;
 	char	c = size & 0xff;
@@ -60,7 +60,7 @@ int
 main(int ac, char **av)
 {
 	int	fd;
-	int	size;
+	size_t	size;
 	int	random = 0;
 	char	*prog = av[0];
